feat(auth): actual working login build_image
