+++ pages/auth/logout.vue
