+++ pages/data-stores/create.vue
@@ -1,9 +1,5 @@
 <script setup lang="ts">
 import ResourceManagerLayout from "~/components/data-stores/ResourceManagerTabs.vue";
-
-definePageMeta({
-  middleware: ["auth"],
-});
 </script>
 
 <template>
