+++ pages/data-stores/index.vue
@@ -1,9 +1,5 @@
 <script setup lang="ts">
 import DataStoreLayout from "~/components/data-stores/DataStoreListTabs.vue";
-
-definePageMeta({
-  middleware: ["auth"],
-});
 </script>
 
 <template>
