+++ pages/proposals.vue
@@ -1,9 +1,5 @@
 <script setup lang="ts">
 import ProposalTable from "~/components/projects/ProposalTable.vue";
-
-definePageMeta({
-  middleware: ["auth"],
-});
 </script>
 
 <template>
