+++ services/auth-service.ts
@@ -27,8 +27,8 @@ export default class AuthService {
         authority: url,
         client_id: clientId,
         client_secret: clientSecret,
-        redirect_uri: `${window.location.origin}/auth`,
-        silent_redirect_uri: `${window.location.origin}/silent-refresh`,
+        redirect_uri: "http://localhost:3000/auth/callback",
+        silent_redirect_uri: `${window.location.origin}/auth/silent-refresh`,
         post_logout_redirect_uri: `${window.location.origin}`,
         response_type: "code",
         userStore: new WebStorageStateStore(),
